Add Shoes trait for Warbler, remove Gemspec
This pulls code out of the Shoes::Swt::Package::Jar and places it
it Warbler::Traits::Shoes. The advantage is that we can now
specify an executable without touching the Warbler code.

Note that the executable chosen by the Shoes trait overrides the
executable chosen by the NoGemspec trait. This may be a coincidence.

The gemspec goes away, too.