Fix bug in JAR creation from within APP
Because the Jar object was created during the App object's
initialization, and  before the APP template was
copied into the 'pkg' directory, the Jar did not know to exclude
the App template files (because they didn't exist when it made its
manifest). As a result, the Jar was bundling the App template, which
added 80MB or so to the Jar size. Now the Jar is only created at
the time when it is needed, so it picks up the fact that extra files
are present, and excludes them.